split out subcomponents
